pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "investment"
version = "1.0-beta"
requires-python = ">=3.12.3"
dependencies = [
    "pandas",
    "typst",
    "requests",
    "PyPDF2",
    "PyYAML"
]

[tool.setuptools.packages.find]
include = ["ai_tool*", "investment*"]

[tool.setuptools.package-data]
"investment.accounting.financialstatements.pdfgeneration" = ["*.typ"]
"investment.tax_report.pdf" = ["*.typ"]

[project.optional-dependencies]
test = [
    "pytest",
]
ai_tool = [
    "openai",
]
